Extract Group from model_soup
Group was the last model class in model_soup, so model_soup is no more.
