+++ grouper/ctl/shell.py
@@ -1,7 +1,6 @@
 import code
 from pprint import pprint
 
-from grouper import model_soup
 from grouper.ctl.util import make_session
 from grouper.graph import GroupGraph
 
@@ -9,7 +8,6 @@ from grouper.graph import GroupGraph
 def shell_command(args):
     session = make_session()
     graph = GroupGraph.from_db(session)
-    m = model_soup
     pp = pprint
 
     try:
@@ -18,8 +16,6 @@ def shell_command(args):
         code.interact(local={
             "session": session,
             "graph": graph,
-            "m": m,
-            "model_soup": model_soup,
             "pp": pp,
         })
     else:
@@ -28,5 +24,5 @@ def shell_command(args):
 
 def add_parser(subparsers):
     shell_parser = subparsers.add_parser(
-        "shell", help="Launch a shell with model_soup imported.")
+        "shell", help="Launch a shell.")
     shell_parser.set_defaults(func=shell_command)
