+++ grouper/fe/handlers/audits_create.py
@@ -7,10 +7,10 @@ from grouper.email_util import send_async_email, send_email
 from grouper.fe.forms import AuditCreateForm
 from grouper.fe.settings import settings
 from grouper.fe.util import GrouperHandler
-from grouper.model_soup import Group
 from grouper.models.audit import Audit
 from grouper.models.audit_log import AuditLog, AuditLogCategory
 from grouper.models.audit_member import AuditMember
+from grouper.models.group import Group
 from grouper.models.group_edge import GROUP_EDGE_ROLES
 from grouper.user_permissions import user_has_permission
