+++ grouper/fe/handlers/group_edit.py
@@ -2,9 +2,9 @@ from sqlalchemy.exc import IntegrityError
 
 from grouper.fe.forms import GroupEditForm
 from grouper.fe.util import GrouperHandler
-from grouper.model_soup import Group
 from grouper.models.audit_log import AuditLog
 from grouper.models.counter import Counter
+from grouper.models.group import Group
 from grouper.service_account import is_service_account
 from grouper.user_group import user_can_manage_group
