+++ grouper/fe/handlers/group_enable.py
@@ -1,6 +1,6 @@
 from grouper.fe.util import GrouperHandler
-from grouper.model_soup import Group
 from grouper.models.audit_log import AuditLog
+from grouper.models.group import Group
 from grouper.service_account import is_service_account
 from grouper.user import user_role
