+++ grouper/fe/handlers/permissions_grant.py
@@ -3,8 +3,8 @@ from sqlalchemy.exc import IntegrityError
 from grouper.audit import assert_controllers_are_auditors, UserNotAuditor
 from grouper.fe.forms import PermissionGrantForm
 from grouper.fe.util import GrouperHandler
-from grouper.model_soup import Group
 from grouper.models.audit_log import AuditLog
+from grouper.models.group import Group
 from grouper.models.permission import Permission
 from grouper.permissions import grant_permission
 from grouper.user_permissions import user_grantable_permissions
