+++ grouper/fe/handlers/public_key_add_tag.py
@@ -1,8 +1,8 @@
 from grouper.fe.forms import PublicKeyAddTagForm
 from grouper.fe.util import GrouperHandler
-from grouper.model_soup import User
 from grouper.models.audit_log import AuditLog
 from grouper.models.public_key_tag import PublicKeyTag
+from grouper.models.user import User
 from grouper.public_key import add_tag_to_public_key, DuplicateTag, get_public_key, KeyNotFound
 from grouper.service_account import can_manage_service_account
 from grouper.user_permissions import user_is_user_admin
