+++ grouper/fe/handlers/search.py
@@ -2,7 +2,7 @@ from sqlalchemy import union_all
 from sqlalchemy.sql import label, literal
 
 from grouper.fe.util import GrouperHandler
-from grouper.model_soup import Group
+from grouper.models.group import Group
 from grouper.models.user import User
