+++ grouper/fe/handlers/user_password_add.py
@@ -2,8 +2,8 @@ from grouper.email_util import send_email
 from grouper.fe.forms import UserPasswordForm
 from grouper.fe.settings import settings
 from grouper.fe.util import GrouperHandler
-from grouper.model_soup import User
 from grouper.models.audit_log import AuditLog
+from grouper.models.user import User
 from grouper.service_account import can_manage_service_account
 from grouper.user_password import add_new_user_password, PasswordAlreadyExists
