+++ grouper/group.py
@@ -1,6 +1,6 @@
 from grouper.graph import Graph, NoSuchGroup
-from grouper.model_soup import Group
 from grouper.models.base.session import Session  # noqa
+from grouper.models.group import Group
 
 
 def get_all_groups(session):
