+++ grouper/service_account.py
@@ -1,9 +1,9 @@
 from collections import namedtuple
 
 from grouper.constants import USER_ADMIN
-from grouper.model_soup import Group
 from grouper.models.audit_log import AuditLog
 from grouper.models.base.session import Session  # noqa
+from grouper.models.group import Group
 from grouper.models.user import User
 from grouper.user import disable_user, enable_user
 from grouper.user_group import user_can_manage_group
