+++ tests/fixtures.py
@@ -1,19 +1,18 @@
 import pytest
 
-from grouper import model_soup
 from grouper.api.routes import HANDLERS as API_HANDLERS
 from grouper.app import Application
 from grouper.constants import AUDIT_MANAGER, PERMISSION_AUDITOR, USER_ADMIN
 from grouper.fe.routes import HANDLERS as FE_HANDLERS
 from grouper.fe.template_util import get_template_env
 from grouper.graph import Graph
-from grouper.model_soup import Group
 from grouper.models.base.model_base import Model
 from grouper.models.base.session import Session, get_db_engine
+from grouper.models.group import Group
+from grouper.models.permission import Permission
 from grouper.models.user import User
 from grouper.permissions import enable_permission_auditing
 from util import add_member, grant_permission
-from grouper.models.permission import Permission
 
 
 @pytest.fixture
@@ -131,7 +130,7 @@ def session(request, tmpdir):
     def fin():
         session.close()
         # Useful if testing against MySQL
-        # model_soup.Model.metadata.drop_all(db_engine)
+        # Model.metadata.drop_all(db_engine)
     request.addfinalizer(fin)
 
     return session
