+++ tests/test_passwords.py
@@ -5,7 +5,7 @@ from tornado.httpclient import HTTPError
 
 from fixtures import standard_graph, graph, users, groups, session, permissions  # noqa
 from fixtures import fe_app as app  # noqa
-from grouper.model_soup import User
+from grouper.models.user import User
 from grouper.user_password import PasswordAlreadyExists, add_new_user_password, delete_user_password, user_passwords
 from url_util import url
